Improve error reporting and remove the need to supply SSH keys (#42)
Improve error reporting in `test-coverage.py`, `release-approval.py`, and `sync-dependencies.py` in the following ways:
- Ensure `stderr` is captured and not discarded which will make debugging much easier.
- Eliminate various scenarios where the script can fail silently when making a failed HTTP request due to incorrect credentials.

`release-approval.py` has been updated to accept Github username and token which can be supplied via the `$RELEASE_APPROVAL_USERNAME` and `$RELEASE_APPROVAL_TOKEN` environment variables. This is an improvement to the old approach of needing SSH keys which is more cumbersome to generate and supply.

Closes https://github.com/graknlabs/build-tools/issues/34 and https://github.com/graknlabs/build-tools/issues/35