Add version interpolation to artifact rules (#154)
We need to be able to publish artifacts with versioned filenames and including help for this logic in the `artifact` rules will significantly improve development experience. The solution we've gone for is to allow the version to be interpolated into the artifact filename, since this can be the same on both sides of the rule (deploy/depend), works in static rules (no file dependencies), is very flexible on filename format and has very clear visibility on what the resulting filename will look like.