Add extract script generation macro (#156)
We need to be able to extract an artifact depended on via artifact rules on the command line in CI. This is tricky without some bazel magic, so we've created a macro that can use bazel specific logic to generate an executable script that performs this logic for us. This plays by all of bazel's rules to avoid non-hermetic builds.