Make checkstyle_test work with any file type (#199)
## What is the goal of this PR?

To make checkstyle_test work with any file type

## What are the changes implemented in this PR?

Deleted `targets` and `files`. Added `include` and `exclude`. They both take in a list of files (e.g. via a `glob`).

If a file is present in `include` it will be checked by checkstyle, unless it is also present in `exclude`. If a file is present in `exclude`, it will not be checked by checkstyle. If a file is present in neither `include` nor `exclude`, an error will be thrown by `checkstyle:test-coverage`.