Add a macro for loading Google OR Tools in a platform independent manner (#216)
## What is the goal of this PR?

Google OR-Tools is not platform-independent - it has separate library files for Mac, Linux and Windows. Therefore, it would be ideal if we can allow the user to load the library from the respective repository without having to deal with making sure it works for the right platform.

The solution we came up with is by adding a Bazel macro for loading the library depending on the platform. For example, if you invoke Bazel in a Mac machine, it will correctly load the library meant for Mac. The same applies to the other operating systems.