Fix create-notes to escapes note value and order most recent items first(#340)
## What is the goal of this PR?

We fix release-note creation by escaping the notes inserted into the release template. The create-notes process used a regex to perform string replacement, which was not treating the replacement value as a literal string, but a regex as well, leading to errors. We also reverse the commits before generating the notes, so we get most-recent-first in the generated github release.

## What are the changes implemented in this PR?

* escape the regex replacement string (ie. the notes markdown)
* reverse the order of commits before generating the release notes, so we get most recent commits first rather than last