Add hash for OR-Tools, remove unneccessary pip repositories call (#372)
## What is the goal of this PR?

We increase hermeticity and reduce the number of warnings output by Bazel by applying recommended fixes with good reason.

Currently, there are a number of warnings output by Bazel upon syncing
dependencies. These four warnings provide some recommendations as to how
we can improve the state of the repository. Additionally, fixing the
problems associated with warnings is good practice as it prevents
developers from becoming used to warnings and potentially missing
something important. Case in point: if someone were to compromise the
google/or-tools repository and replace the releases we download with
compromised packages, Bazel wouldn't know.

We are currently missing a hash for our `http_archive` downloads for
`google_or_tools_*`. This reduces hermeticity.

We currently call pip_repositories, despite the fact that it is a
deprecated method.

## What are the changes implemented in this PR?

- We have added a sha256 argument to all of our `google_or_tools_*`
`http_archive` calls, increasing hermeticity.
- We have removed our call to `pip_repositories`.