Rust IDE sync tool (#375)
## What is the goal of this PR?

We've added tooling to "sync" a Bazel project that uses Rust via
`@rules_rust` such that IDEs, like IntelliJ and CLion, are able to
support its development - providing IntelliSense and so on.

## What are the changes implemented in this PR?

When we began work on the IntelliJ Bazel + Rust plugins 9 months ago, we
aimed to create fully-functional IDE tooling that was fully faithful to
Bazel's methodology. Due to the substantial difficulty of getting things
to work (e.g: unable to debug a Bazel **rust_test**), we changed
strategy. Taking bits and pieces of the work we'd done on those plugins
(most notably the Bazel Plugin's `Cargo.toml` generation), we no longer
centre our development workflow around Bazel targets, but rather add
tooling to generate Cargo projects from our Bazel projects
automatically. This allows us to leverage the full power of the IntelliJ
Rust plugin to develop seamlessly in Rust.

### How it works

We added a Kotlin binary (`@vaticle_dependencies//ide/rust:sync`) that,
when run in a Bazel project that uses `rules_rust`, generates the
`Cargo.toml` files that define a Cargo project. It runs the following
steps:

1. Validate the project Rust targets by doing `bazel build
--build=false` on the targets whose names start with `rust_`. This
verifies correctness of the Rust `BUILD` targets, as things like
incorrect dependency declarations can derail the sync tool.
2. In the Bazel workspace directory, `bazel build` the Rust targets.
This achieves two purposes: firstly, it loads the Rust toolchain and
stdlib that IntelliJ and CLion use; secondly, it loads the external
`@vaticle` Rust repos so we can sync them too. Errors at this stage are
ignored, as they typically represent errors in our Rust code, and we
want the user to be able to sync the project in that scenario.
3. In the current workspace, and in all external `@vaticle`
repositories* that Rust targets depend on:
1. Run the **Rust IDE Sync Info Aspect**, outputting metadata in the
form of `.ide-sync.properties` files in `bazel-bin`. Using an Aspect
allows us to easily retrieve crate names and versions of the entire
dependency tree.
2. Run the **Cargo Project Generator**, a Kotlin/JVM binary that
recursively searches `bazel-bin` for IDE metadata, and transforms them
to `Cargo.toml` files in the project directory. The Aspect alone can't
do this, because a) it can only write to `bazel-bin` and b) Cargo and
Bazel differ in the way test dependencies are declared; a Cargo library
needs to know about its tests, but a Bazel Rust test target needs to
know about its library.

\* _"Repository" here refers to a [Bazel
repository](https://bazel.build/extending/repo), such as
`@vaticle_typeql`._

#### Why sync external `@vaticle` repositories?

Whenever a Rust target in our current project depends on an external
`@vaticle` Rust target, its generated `Cargo.toml` contains a reference
to that external target. Any dependency in a `Cargo.toml` must itself
have a `Cargo.toml`; hence, we must sync all external `@vaticle`
dependencies as well as the current project.

Why use "starts with `@vaticle`" as the filter, rather than something
more generic? Because it keeps the logic very simple. If, in the future,
we depend on non-`vaticle` Bazel Rust projects, we'll need to update the
filter.

### Usage

In each Bazel project that uses Rust, add the following executable
script, which we name `dependencies/ide/sync.sh`:
```sh
bazel run @vaticle_dependencies//ide/rust:sync
```
After "importing the Bazel project", immediately run
`dependencies/ide/sync.sh`. Once `sync.sh` completes, the Cargo project
will be generated.

Add `target`, `Cargo.lock` and `Cargo.toml` to `.gitignore`. (`target`
is generated by a Cargo build.)

The Sync tool supports the `--verbose` (`-v`) command-line flag to
enable verbose output (`bazel run @vaticle_dependencies//ide/rust:sync
-- --verbose`)

### Compatibility requirements

Currently, the sync tool only works with `@vaticle` Bazel dependencies
(see "Why sync external `@vaticle` repositories" above). If we want to
depend on Bazel projects from a different org, we'll need to update the
relevant code.

We also assume that the Rust support in Bazel is delivered via
`@rules_rust`. This is a fair assumption; it is the standard library for
Rust rules.

We assume that external crates are loaded via `cargo raze`. This is
currently true for all crates declared in `@vaticle_dependencies`. If we
were to refactor in the future to use, say, `crate_universe`, we would
inevitably have to update the sync tool as well.

It's also possible that API changes and refactors in `rules_rust` will
result in further changes needing to be made to our sync tool. For
example, a build script is currently detected on the basis of its label
being `_build_script_run`. As this is an internal/private symbol in
`rules_rust`, it's quite likely to change in the future.