Fix IDE sync tool generating binary names that are illegal in Cargo.toml (#378)
## What is the goal of this PR?

We fixed the IDE sync tool generating binary names that are illegal in
`Cargo.toml`, and made the sync tool have `public` visibility so that
other repos can verify it.

## What are the changes implemented in this PR?

Generated binary names are based on their project-relative path,
ensuring uniqueness. But, they contained slashes, which are illegal in a
`Cargo.toml` `[bin.name]`. So we now replace them with underscores.

Also, we've made the sync tool `public` so that other repos can declare
it in their `:ci` `BUILD` targets.