Add transitive deps of `ortools` that aren't included by default on M1 Macs (#386)
## What is the goal of this PR?

We added `core-foundation-sys` and `libc` explicitly as crate
dependencies. They are conditional dependencies of `ortools`. Due to a
bug, most likely in `cargo raze`, they are not included by default.

## What are the changes implemented in this PR?

`core-foundation-sys` and `libc` are conditional dependencies of the
`ortools` crate. The condition is the Cargo condition `cfg('unix')`,
which is supposed to represent any Unix system. Due to a bug, most
likely in `cargo raze`, they are not correctly pulled in when the
machine has `aarch64` architecture.

We'll raise an issue/PR in the relevant upstream repo (most likely
`cargo-raze`), but for now, we need a workaround to get our projects
building on an M1 Mac. This PR is that workaround. We'll declare the
dependencies explicitly, and depend on them explicitly in the targets
that require them - most notably `ortools` will explicitly depend on
`libc`, and `typeql` will explicitly depend on `libc` and
`core-foundation-sys`.