IDE sync: generate correct paths for crates containing generated sources (#398)
## What is the goal of this PR?

The Rust IDE sync tool now generates correct paths for crates that
contain generated code.

## What are the changes implemented in this PR?

The IDE sync tool had a "sources are generated" flag which triggered
when the entry point was generated, but not when other files (like
TypeDB Protocol's `typedb.protocol.rs`) were generated. In this PR, we
change "sources are generated" to "contains generated sources". When a
crate contains generated sources, its generated `Cargo.toml` will now
have its entry point set to a location inside `bazel-bin`.

### Why does this work?

The sync tool builds the project's Rust targets. When a target contains
generated sources, Bazel not only places the generated sources in
`bazel-bin`, but also creates a symlink in `bazel-bin` for every
non-generated source file that links to the source file in the project's
working directory. The above behaviour was added in:
- https://github.com/bazelbuild/rules_rust/pull/1340

So as long as the sync tool generates a `Cargo.toml` that points to
`bazel-bin`, the resulting `Cargo.toml` can be successfully built using
`cargo`.