Rust IDE sync: compile generated sources (#400)
## What is the goal of this PR?

The Rust IDE sync tool now compiles generated sources in external
repositories.

## What are the changes implemented in this PR?

Previously, `typedb-client-rust` would fail to build using `cargo`,
because the sources of `typedb-protocol` - imported as an external Bazel
repository - were not automatically compiled by the sync tool. So the
generated `Cargo.toml` for that target pointed to a valid `lib.rs` file,
but that `lib.rs` file depended on `typedb.protocol.rs`, which did not
exist, because it had not been created.

Now, whenever the sync tool finds a Bazel Rust target that contains
generated sources, it runs `bazel build` on that target to generate
those sources.