Don't use Bazel in crate-universe lockfile update script (#411)
## What is the goal of this PR?

rules_rust's own way of repinning changed dependencies in
`crates_repository`, i.e. by setting the `CARGO_BAZEL_REPIN` environment
variable to `full` takes a prohibitively long time for minor changes to
the cargo manifest. We instead opted for using `cargo
generate-lockfile`, which takes an order of magnitude less time to
complete (~10 s as opposed to 3-5 minutes).

Using the cargo executable provided by rules_rust seems to be the most
correct approach to updating the lockfile. However, calling it directly
requires Bazel to build the workspace. Since we have rust executables in
dependencies (e.g. `//builder/grpc/rust:compile`), Bazel will attempt to
fetch the crates. Fetching the crates will fail if the manifest and the
lockfile don't match, which they won't, since we just changed the
manifest and are attempting to regenerate the lockfile.

The options are:
1. use `CARGO_BAZEL_REPIN=full`, slow as it is;
2. use system `cargo`;
3. edit WORKSPACE temporarily from within `update.sh`;
4. download a `cargo` binary and use that to generate the lockfile.

We've chosen option 4 for this PR, as that is fast, doesn't rely on the
user system as much as option 2, and isn't as hacky as option 3. It is,
however, limited to darwin-x64 at present.

## What are the changes implemented in this PR?

In `library/crates/update.sh`, fetch the cargo executable from
repo.vaticle.com and regenerate the lockfile using that.