Bazel Cache Installation Tool (#433)
## What is the goal of this PR?

We've updated the Bazel cache installation tool, bringing it up to date with our latest approach.

We will run this file before every Factory job that wants to make use of the Bazel cache. This will not interfere with development or use by developers outside of Vaticle. Developers inside Vaticle will be able to make use of the cache by running this tool after exporting the appropriate environment variables. 

Repositories where this workflow will be common should add '.bazel-remote-cache.rc' and 'credential.json to .gitignore.

## What are the changes implemented in this PR?

The tool now checks for the presence of non-empty $BAZEL_CACHE_CREDENTIAL and $BAZEL_CACHE_URL environment variables. If they are both found, we generate a 'bazel_cache_credential.json' file from the content of $BAZEL_CACHE_CREDENTIAL, which is expected to be a base64 encoded string of a JSON GCP credential file. If it does exist, but fails to deserialise, we exit with an error code 1.