Introduce Cargo project aspect, Cbindgen & SWIG rules (#447)
We introduce the cargo project aspect, which allows us to convert a
rust_library target into a cargo workspace within the Bazel sandbox.
This allows us to run any tools that expect a Cargo project as an input,
such as the C bindings generator Cbindgen.

We also add SWIG as a Bazel dependency and provide the Java wrapper
rules for it.

- We convert the Cargo sync properties aspect into a Rust Cargo project
aspect, which generates a full Cargo workspace in the Bazel sandbox,
including external dependencies and generated sources, for each Rust
target.
- We modify the Rust IDE sync tool so that it uses the generated project
in the sandbox as the non-crates.io external dependencies.
- We add cbindgen as a crates_universe dependency and provide the
associated Bazel rule.
- We build SWIG from source and provide a set of rules that allow us to
wrap a C library into a Java module.