Fix checkstyle coverage check for filenames with whitespaces (#478)
## What is the goal of this PR?

The checkstyle coverage check failed if there were files in the workspace with whitespaces in filenames. We fix it.

## What are the changes implemented in this PR?

The list of files was generated by splitting the `find` command result by all whitespace characters. But each file is listed in a separate line, that allows us to split by `"\n"` only.