Add a universal aspect_bazel_lib dependency for csharp and nodejs rules (#547)
`typedb-driver` and `typedb-protocol` builds have been broken after a
recent `ts` rules
[update](https://github.com/vaticle/dependencies/commit/73f3514) because
of the updated `aspect_bazel_lib` dependency.

To fix the current issue and prevent its reproduction, we add a
universal `aspect_bazel_lib` dependency with a specific version, which
is maximal for all the dependent rules in `dependencies`. This
dependency should be explicitly declared before using any rule that
depends on `aspect_bazel_lib` of any version and updated every time a
rule uses a newer version of it as its dependency.