Add integration tests to generated Cargo.toml (#549)
## What is the goal of this PR?

We now generate a `Cargo.toml` entry per Bazel `rust_test` entry, if
that test is a module/crate on its own and not a set of unit tests, that
apply for a specific existing crate (eg. using the `crate = ":target"`
attribute).

We support the conversion of Bazel Rust integration test targets when
there is exactly 1 source file in the `rust_test` target. Otherwise, we
expect that the `crate_root` be set to the source file that contains the
test targets.

For example:
```
rust_test(
  name = "test_abc",
  srcs = ["test_abc_v_1.rs, "common/utils.rs"]
)
```
is now considered ambiguous, and must be updated to:
```
rust_test(
  name = "test_abc",
  crate_root = "test_abc_v_1.rs",
  srcs = ["test_abc_v_1.rs, "common/utils.rs"]
)
```

Unrelatedly, we add `resolver = "2"` for into the root `Cargo.toml`,
whenever we need to generate a Cargo workspace.

## What are the changes implemented in this PR?

* The rust project aspect now generates an `entry_point_path` property
for test targets, as well as binaries and libraries, if the test is a
integration test (eg no `crate = :target` set).
* The cargo toml generator adds test entries as well a benchmark entries
for integration tests in the crate/package