Adjust cargo IDE sync tool to depend on external dependencies via git (#560)
## What is the goal of this PR?

As part of the initiative to streamline the Rust developer experience,
we aim to commit autogenerated cargo manifests. For them to be usable
without Bazel, we add the ability to depend on internal git repositories
rather than pointing the dependency via path into the generated cargo
workspace.

## What are the changes implemented in this PR?

We also merge crates at the same path that have the same name under the
assumption that they represent alternative modes of compilation for the
same target (i.e., enable different features).