Add dependencies used in diagnostics package of the Rust TypeDB server (#564)
## Usage and product changes
We add Rust crate dependencies for the TypeDB Server that are used by
the newly introduced diagnostics package:
* [hyper](https://crates.io/crates/hyper) - one of the most popular HTTP
crates used for working with REST API (creating HTTP clients and
servers);
* [hyper-rustls](https://crates.io/crates/hyper-rustls) - TLS (HTTPS)
support for `hyper` without dependencies on `openssl`;
* [sysinfo](https://crates.io/crates/sysinfo) - a crate for retrieving
information about the system (memory, disks, processes, etc.);
* [xxhash-rust](https://crates.io/crates/xxhash-rust) - an
implementation for a hasher with consistent results over all OSs, archs,
and server reboots (used for database names hashing).

We modify the `RustManifestSyncer` so it can process source (and compile
time) files from a non-target (sub)directory. The original version used
to abort the sync tool process in this case, blocking, for example,
adding `VERSION` files from the root of a project as compile data to a
target.