Add dependencies for rust sentry crate (#565)
## Usage and product changes
Add dependencies for [rust sentry
crate](https://crates.io/crates/sentry/0.36.0) to allow sending critical
error events from TypeDB Server.

## Motivation


## Implementation
The following dependencies are affected:
* `sentry` is added with the `ureq` feature enabled for events transport
serving. The default `reqwest` dependency is not acceptable as it
introduces an `openssl` dependency to the server (even if we enable the
`rustls` feature), while `ureq` does not require any additional
libraries to be set on the host (unlike `curl` which requires `libcurl`
installation).
* `ureq` is explicitly added to enable the `tls` feature.
* `async-stripe`'s version is upgraded to eliminate version conflicts in
common dependencies with `sentry` (there is an assumption that it's no
longer used in our repos and can be freely updated / eliminated).